requirements.txt
pyqt6
requests
pillow
qdarktheme
